pr clean ups and move supabase createclient to common/data/supabase.
