+++ src/common/data/config/appConfig.ts
@@ -5,4 +5,4 @@ export const NEXT_PUBLIC_SUPABASE_ANON_KEY =
 export const NEXT_PUBLIC_SUPABASE_URL = process.env.NEXT_PUBLIC_SUPABASE_URL
 export const SUPABASE_SERVICE_ROLE_KEY = process.env.SUPABASE_SERVICE_ROLE_KEY
 
-export const HOME_PAGE = '/candidate/job-tracker'
+export const CANDIDATE_HOME_PAGE = '/candidate/job-tracker'
